Rename first.c to cacheSim.c
Cache Simulator I made for computer architecture. The only error this really seemed to have was that at size 32, assoc:4 my hits and misses for the non-prefetch were off by one for larger files. Otherwise prefetch was working fine and the non-prefetch was ok for other test cases as well.